requirements.txt
google-auth
requests
requests-toolbelt
pyopenssl
ndg-httpsclient
pyasn1
pillow
firebase-admin
google-api-python-client